Split extract MiniSatSolver from MiniSat
Everything related to search in MiniSat is useless for LCG.
So, a subclass can be created only for that purpose, namely MiniSatSolver.

+ activate topLevelCleanUp in LCG
