+++ solver/src/main/java/org/chocosolver/sat/MiniSat.java
@@ -10,17 +10,13 @@
 package org.chocosolver.sat;
 
 import gnu.trove.list.TIntList;
-import gnu.trove.list.array.TDoubleArrayList;
 import gnu.trove.list.array.TIntArrayList;
 import gnu.trove.map.hash.TIntObjectHashMap;
 import org.chocosolver.solver.variables.impl.LitVar;
-import org.chocosolver.util.ESat;
-import org.chocosolver.util.objects.IntHeap;
 
 import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.Comparator;
-import java.util.Random;
 
 /**
  * <p>A MiniSat solver.</p>
@@ -41,13 +37,14 @@ import java.util.Random;
  * @author Charles Prud'homme
  * @since 12/07/13
  */
-public class MiniSat implements SatFactory, Dimacs {
+@SuppressWarnings("FieldCanBeLocal")
+public class MiniSat implements SatFactory {
 
-    private static final int DEBUG = 0;
+    static final int DEBUG = 0;
     // Value of an undefined variable
-    private static final int varUndef = -1;
+    static final int varUndef = -1;
     // value of an undefined literal
-    private static final int litUndef = -2;
+    static final int litUndef = -2;
 
     public static final int lTrue = 0b01;
     public static final int lFalse = 0b10;
@@ -55,7 +52,7 @@ public class MiniSat implements SatFactory, Dimacs {
     // undefined clause
     protected static ThreadLocal<Integer> clauseCounter = ThreadLocal.withInitial(() -> 0);
     public static final Clause C_Undef = Clause.undef();
-    private static final Reason R_Undef = Reason.undef();
+    static final Reason R_Undef = Reason.undef();
     static final VarData VD_Undef = new VarData(R_Undef, -1, -1);
     public Clause confl = C_Undef;
 
@@ -68,65 +65,41 @@ public class MiniSat implements SatFactory, Dimacs {
     // List of problem addClauses.
     public final ArrayList<Clause> clauses = new ArrayList<>();
     // List of learnt addClauses.
-    private final ArrayList<Clause> learnts = new ArrayList<>();
+    final ArrayList<Clause> learnts = new ArrayList<>();
     // 'watches_[lit]' is a list of constraints watching 'lit'(will go
     // there if literal becomes true).
     private final TIntObjectHashMap<ArrayList<Watcher>> watches_ = new TIntObjectHashMap<>();
     // The current assignments.
-    //TIntObjectHashMap<Boolean> assignment_ = new TIntObjectHashMap<>();
-    TIntArrayList assignment_ = new TIntArrayList();
+    final TIntArrayList assignment_ = new TIntArrayList();
     // Assignment stack; stores all assignments made in the order they
     // were made.
-    TIntArrayList trail_ = new TIntArrayList();
+    final TIntArrayList trail_ = new TIntArrayList();
     // Separator indices for different decision levels in 'trail_'.
-    TIntArrayList trail_markers_ = new TIntArrayList();
+    final TIntArrayList trail_markers_ = new TIntArrayList();
     // Head of queue(as index into the trail_).
     int qhead_;
     // Number of variables
     int num_vars_;
     int rootlvl = 0;
-    int ccmin_mode = 0; // Controls conflict clause minimization (0=none, 1=basic, 2=deep)
-    int phase_saving = 0; // Controls the level of phase saving (0=none, 1=limited, 2=full)
-    double cla_inc = 1;
-    double var_inc = 1;
-    double var_decay = 0.95;
-    double clause_decay = 0.999;
-    double random_var_freq = 0;
-    int restart_inc = 2;
-    boolean rnd_init_act = true;
-    boolean luby_restart = true;
-    int restart_first = 100;
-    int random_seed = 7;
+    private final int ccmin_mode = 0; // Controls conflict clause minimization (0=none, 1=basic, 2=deep)
+    private double cla_inc = 1;
+    private final double clause_decay = 0.999;
+
     double learntsize_adjust_confl = 100;
     int learntsize_adjust_cnt = 100;
-    double learntsize_adjust_inc = 1.5;
-    double learntsize_inc = 1.1;
-    double learntsize_factor = 1 / 3d;
-    boolean rnd_pol;
-    int conflict_budget = -1;
-    int propagation_budget = -1;
+    final double learntsize_adjust_inc = 1.5;
+    final double learntsize_inc = 1.1;
+    final double learntsize_factor = 1 / 3d;
     int propagations;
-    int rnd_decisions;
-    boolean asynch_interrupt = false;
-    TIntArrayList model = new TIntArrayList();
-    TIntArrayList conflict = new TIntArrayList();
-    ArrayList<VarData> vardata = new ArrayList<>();
-    ArrayList<ChannelInfo> cinfo = new ArrayList<>();
-    int conflicts;
-    int decisions;
-    int max_literals;
-    int tot_literals;
-    int dec_vars;
-    int clauses_literals;
-    int learnts_literals;
+    final ArrayList<VarData> vardata = new ArrayList<>();
+    final ArrayList<ChannelInfo> cinfo = new ArrayList<>();
+    private int max_literals;
+    private int tot_literals;
+    private int clauses_literals;
+    private int learnts_literals;
     double max_learnts;
-    BitSet seen = new BitSet();
-    BitSet decision = new BitSet();
-    BitSet polarity = new BitSet();
-    TIntArrayList analyze_toclear = new TIntArrayList();
-    TDoubleArrayList activity = new TDoubleArrayList();
-    IntHeap order_heap = new IntHeap((a, b) -> activity.get(a) > activity.get(b));
-    Random rand;
+    final BitSet seen = new BitSet();
+    private final TIntArrayList analyze_toclear = new TIntArrayList();
     private final TIntArrayList temporary_add_vector_ = new TIntArrayList();
 
     /**
@@ -136,7 +109,6 @@ public class MiniSat implements SatFactory, Dimacs {
         this.ok_ = true;
         this.qhead_ = 0;
         num_vars_ = 0;
-        rand = new Random(random_seed);
         assignment_.add(lUndef); // required because variable are numbered from 1
         if (addTautology) {
             // true literal
@@ -183,21 +155,10 @@ public class MiniSat implements SatFactory, Dimacs {
         assignment_.add(lUndef);
         vardata.add(VD_Undef);
         cinfo.add(ci);
-        //activity .push(0);
-        activity.add(rnd_init_act ? rand.nextDouble() * 0.00001 : 0);
         seen.clear(v);
-        polarity.set(v);
-        if (!decision.get(v)) dec_vars++;
-        decision.set(v);
-        insertVarOrder(v);
         return v;
     }
 
-    private void insertVarOrder(int v) {
-        if (!order_heap.contains(v) && decision.get(v)) {
-            order_heap.insert(v);
-        }
-    }
 
     public void beforeAddingClauses() {
         // nothing to do by default.
@@ -296,7 +257,7 @@ public class MiniSat implements SatFactory, Dimacs {
 
     public void addLearnt(TIntList learnt_clause) {
         for (int v = 0; v < nVars(); v++) {
-            assert valueVar(v) != MiniSat.lUndef || order_heap.contains(v) : v + " not heaped";
+            assert valueVar(v) != MiniSat.lUndef : v + " not heaped";
         }
         if (learnt_clause.size() == 1) {
             uncheckedEnqueue(learnt_clause.get(0));
@@ -307,7 +268,6 @@ public class MiniSat implements SatFactory, Dimacs {
             claBumpActivity(cr);
             uncheckedEnqueue(learnt_clause.get(0), Reason.r(cr));
         }
-        varDecayActivity();
         claDecayActivity();
 
         if (--learntsize_adjust_cnt == 0) {
@@ -335,9 +295,6 @@ public class MiniSat implements SatFactory, Dimacs {
                     else
                         System.out.printf("Unfix %d\n", trail_.get(c));
                 }
-                if (phase_saving > 1 || (phase_saving == 1) && c > trail_markers_.get(trail_markers_.size() - 1))
-                    polarity.set(x, sgn(trail_.get(c)));
-                insertVarOrder(x);
             }
             qhead_ = trail_markers_.get(level);
             trail_.remove(trail_markers_.get(level), trail_.size() - trail_markers_.get(level));
@@ -405,7 +362,7 @@ public class MiniSat implements SatFactory, Dimacs {
     }
 
 
-    private int incrementVariableCounter() {
+    int incrementVariableCounter() {
         return num_vars_++;
     }
 
@@ -614,138 +571,6 @@ public class MiniSat implements SatFactory, Dimacs {
         return false;
     }
 
-
-    /**
-     * A call to this method will attempt to find
-     * an interpretation that satisfies the Boolean formula declared in this.
-     *
-     * @return {@code ESat.TRUE} if such an interpretation is found,
-     * {@code ESat.FALSE} if no interpretation exists,
-     * {@code ESat.UNDEFINED} if a limit was reached.
-     */
-    public ESat solve() {
-        model.clear();
-        conflict.clear();
-        if (!ok_) return ESat.FALSE;
-        max_learnts = nClauses() * learntsize_factor;
-        learntsize_adjust_confl = 100;
-        learntsize_adjust_cnt = (int) learntsize_adjust_confl;
-        ESat status = ESat.UNDEFINED;
-
-        // Search:
-        int curr_restarts = 0;
-        while (status == ESat.UNDEFINED) {
-            double rest_base = luby_restart ? luby(restart_inc, curr_restarts) : Math.pow(restart_inc, curr_restarts);
-            status = search((int) (rest_base * restart_first));
-            if (!withinBudget()) break;
-            curr_restarts++;
-        }
-
-        if (status == ESat.TRUE) {
-            // Extend & copy model:
-            model.ensureCapacity(nVars());
-            for (int i = 0; i < nVars(); i++) {
-                model.add(valueLit(i));
-            }
-
-        } else if (status == ESat.FALSE && conflict.isEmpty())
-            ok_ = false;
-
-        cancelUntil(0);
-        if (status == ESat.TRUE) {
-            System.out.print("SAT\n");
-            for (int i = 0; i < nVars(); i++)
-                if (model.get(i) != lUndef)
-                    System.out.printf("%s%s%d", (i == 0) ? "" : " ",
-                            (model.get(i) == lTrue) ? "" : "-", i + 1);
-            System.out.print(" 0\n");
-        } else if (status == ESat.FALSE)
-            System.out.print("UNSAT\n");
-        else
-            System.out.print("INDET\n");
-        return status;
-    }
-
-    /**
-     * Search for a model the specified number of conflicts.
-     *
-     * @param nof_conflicts limit over the number of conflicts
-     * @implNote Use negative value for 'nof_conflicts' indicate infinity.
-     */
-    ESat search(int nof_conflicts) {
-        assert ok_;
-        int backtrack_level;
-        int conflictC = 0;
-        TIntArrayList learnt_clause = new TIntArrayList();
-
-        for (; ; ) {
-            propagate();
-            if (confl != C_Undef) {
-                // CONFLICT
-                conflicts++;
-                conflictC++;
-                if (trailMarker() == 0) return ESat.FALSE;
-
-                learnt_clause.clear();
-                backtrack_level = analyze(confl, learnt_clause);
-                cancelUntil(backtrack_level);
-                addLearnt(learnt_clause);
-
-            } else {
-                // NO CONFLICT
-                if (nof_conflicts >= 0 && conflictC >= nof_conflicts || !withinBudget()) {
-                    // Reached bound on number of conflicts:
-                    cancelUntil(0);
-                    return ESat.UNDEFINED;
-                }
-
-                // Simplify the set of problem clauses:
-                if (trailMarker() == 0 && !simplify())
-                    return ESat.FALSE;
-
-                if (learnts.size() - trail_.size() >= max_learnts)
-                    doReduceDB();
-
-                // New variable decision:
-                decisions++;
-                int next = pickBranchLit();
-
-                if (next == litUndef)
-                    // Model found:
-                    return ESat.TRUE;
-
-                // Increase decision level and enqueue 'next'
-                pushTrailMarker();
-                uncheckedEnqueue(next, R_Undef);
-            }
-        }
-
-    }
-
-    int pickBranchLit() {
-        int next = varUndef;
-
-        // Random decision:
-        if (rand.nextDouble() < random_var_freq && !order_heap.isEmpty()) {
-            next = order_heap.get(rand.nextInt(order_heap.size()));
-            if (valueVar(next) == lUndef && decision.get(next))
-                rnd_decisions++;
-        }
-
-        // Activity based decision:
-        while (next == varUndef || valueVar(next) != lUndef || !decision.get(next))
-            if (order_heap.isEmpty()) {
-                next = varUndef;
-                break;
-            } else {
-                next = order_heap.removeMin();
-            }
-
-        return next == varUndef ?
-                litUndef :
-                makeLiteral(next, rnd_pol ? rand.nextDouble() < 0.5 : polarity.get(next));
-    }
-
     public int findConflictLevel() {
         int lvl = -1;
         for (int i = 0; i < confl.size(); i++) {
@@ -800,7 +625,7 @@ public class MiniSat implements SatFactory, Dimacs {
             j = out_learnt.size();
 
         max_literals += out_learnt.size();
-        out_learnt.subList(j, out_learnt.size()).clear();
+        out_learnt.remove(j, out_learnt.size() - j);
         tot_literals += out_learnt.size();
 
         // Find correct backtrack level:
@@ -837,7 +662,7 @@ public class MiniSat implements SatFactory, Dimacs {
                 int x = var(q);
                 if (!seen.get(x) && level(x) > rootlvl) {
                     assert p == litUndef || pos(var(p)) > pos(x) : "chronological inconsistency :(" + printLit(p) + " @ " + pos(var(p)) +
-                            ") is explained by a previous event (" + printLit(x) + " @ " + pos(x) + ") "+c;
+                            ") is explained by a previous event (" + printLit(x) + " @ " + pos(x) + ") " + c;
                     varBumpActivity(x);
                     seen.set(x);
                     if (DEBUG > 1) System.out.printf("mark %d\n", x);
@@ -922,35 +747,45 @@ public class MiniSat implements SatFactory, Dimacs {
 
     boolean simplify() {
         assert (trailMarker() == rootlvl);
-
         if (ok_) propagate();
         if (!ok_ || (confl != C_Undef))
             return ok_ = false;
         // TODO
-        /*if (nAssigns() == simpDB_assigns || (simpDB_props > 0))
-            return true;
+//        if (nAssigns() == simpDB_assigns || (simpDB_props > 0))
+//            return true;
 
         // Remove satisfied clauses:
         removeSatisfied(learnts);
-        if (remove_satisfied)        // Can be turned off.
-            removeSatisfied(clauses);
-        */
-        rebuildOrderHeap();
+//        if (remove_satisfied)        // Can be turned off.
+//            removeSatisfied(clauses);
 
-        /*simpDB_assigns = nAssigns();
-        simpDB_props = clauses_literals + learnts_literals;   // (shouldn't depend on stats really, but it will do for now)
-        */
         return true;
     }
 
-    private void rebuildOrderHeap() {
-        TIntList vs = new TIntArrayList();
-        for (int v = 0; v < nVars(); v++)
-            if (decision.get(v) && valueVar(v) == lUndef)
-                vs.add(v);
-        order_heap.build(vs);
+    void removeSatisfied(ArrayList<Clause> cs) {
+        int i, j;
+        for (i = j = 0; i < cs.size(); i++) {
+            Clause c = cs.get(i);
+            if (satisfied(c)){
+                removeClause(cs.get(i));
+            }else{
+                cs.set(j++, cs.get(i));
+
+            }
+        }
+//        cs.shrink(i - j);
+        if (cs.size() > j) {
+            cs.subList(j, cs.size()).clear();
+        }
     }
 
+    boolean satisfied(Clause c) {
+        for (int i = 0; i < c.size(); i++)
+            if (valueLit(c._g(i)) == lTrue) {
+            return true;
+        }
+        return false;
+    }
 
     public void doReduceDB() {
         int i, j;
@@ -969,13 +804,6 @@ public class MiniSat implements SatFactory, Dimacs {
         int n = learnts.size();
         learnts.subList(j, n).clear();
         // System.out.printf("reduceDB removed %d clauses\n", n - j);
-    }
-
-
-    boolean withinBudget() {
-        return !asynch_interrupt &&
-                (conflict_budget < 0 || conflicts < conflict_budget) &&
-                (propagation_budget < 0 || propagations < propagation_budget);
     }
 
     Clause getConfl(int p) {
@@ -1022,49 +850,12 @@ public class MiniSat implements SatFactory, Dimacs {
     }
 
     void varBumpActivity(int v) {
-        varBumpActivity(v, var_inc);
+        // empty
     }
 
-    void varBumpActivity(int v, double inc) {
-        activity.ensureCapacity(nVars());
-        double a = activity.get(v);
-        activity.setQuick(v, a + inc);
-        if (a + inc > 1e100) {
-            activity.transformValues(value -> value * 1e-100);
-            var_inc *= 1e-100;
-        }
-        // Update order_heap with respect to new activity:
-        if (order_heap.contains(v))
-            order_heap.decrease(v);
-    }
-
-    void varDecayActivity() {
-        var_inc *= (1 / var_decay);
-    }
 
     void claDecayActivity() {
         cla_inc *= (1 / clause_decay);
-    }
-
-
-    private static double luby(double y, int x) {
-
-        // Find the finite subsequence that contains index 'x', and the
-        // size of that subsequence:
-        int size = 1;
-        int seq = 0;
-        while (size < x + 1) {
-            seq++;
-            size = 2 * size + 1;
-        }
-
-        while (size - 1 != x) {
-            size = (size - 1) >> 1;
-            seq--;
-            x = x % size;
-        }
-
-        return Math.pow(y, seq);
     }
 
     ///////
@@ -1220,13 +1011,13 @@ public class MiniSat implements SatFactory, Dimacs {
             this.pos = pos;
         }
 
-        private void set(Reason cr, int level, int pos){
+        private void set(Reason cr, int level, int pos) {
             this.cr = cr;
             this.level = level;
             this.pos = pos;
         }
 
-        private void clearReason(){
+        private void clearReason() {
             this.cr = R_Undef;
         }
     }
