+++ solver/src/main/java/org/chocosolver/solver/search/loop/learn/LazyClauseGeneration.java
@@ -100,7 +100,7 @@ public class LazyClauseGeneration implements Learn {
             }
         } else {
             nbRestarts = mSolver.getRestartCount();
-            if(mSolver.getSearchWorldIndex() == mSolver.getEnvironment().getWorldIndex()){
+            if(mSolver.getSearchWorldIndex() == mSolver.getEnvironment().getWorldIndex() -1){ // TODO check condition
                 mSat.topLevelCleanUp();
             }
         }
