+++ solver/src/test/java/org/chocosolver/sat/SatTest.java
@@ -32,7 +32,7 @@ public class SatTest {
 
     @Test(groups = "1s")
     public void test1() {
-        MiniSat sat = new MiniSat(false);
+        MiniSatSolver sat = new MiniSatSolver();
         int a = sat.newVariable();
         int b = sat.newVariable();
         sat.addClause(MiniSat.makeLiteral(a), MiniSat.neg(MiniSat.makeLiteral(b)));
@@ -76,7 +76,7 @@ public class SatTest {
 
     private void run(String path, boolean sat) throws FileNotFoundException {
         String file = this.getClass().getResource(path).getFile();
-        MiniSat solver = new MiniSat(false);
+        MiniSatSolver solver = new MiniSatSolver();
         solver.parse(file);
         ESat ret = solver.solve();
         Assert.assertEquals(ret, sat ? ESat.TRUE : ESat.FALSE, "Unexpected search state");
